Add first calculator to the Planner
I added the Max Payment calculator to the Planner component. I'm keeping
the results stored in the App component, since I want to use those
values in other places. The Planner component will use useEffect to
update the results state everytime a value inside the calculator is
modified. I get a lint warning that I chose to ignore. I added the
updateResults function as a dependency in useEffect, but then it kept
running infinitely.

I'm thinking about breaking the individual calcualtors out into their
own components so Planner won't have as much in it. Not set on that idea
yet though.